Random Useless Poop
I expanded on the text readouts and added dumb placeholder values. I
also removed the input parameter as right now its just kind of there and
we wont need it for a while or possibly at all. If we're only showing
the life of a single star type (one solar mass) then there's only one
potential value and thus no need for an input. We can add it back in
later if we make it that far.