More Random Pewp
Messed with some of the placeholder values, put the initialization
commands into one cell for simplicity, made 300 nested If statements.

We should start getting to work on building a database now, as we finish
up the interface. I'm thinking we have a series of files with the same
layout, with the data for one mass range in each file. This would let us
integrate mass types easily, because we would just make a new file for
each mass range.