First Progress Report
I've basically just thrown together the things we need for our progress
report (due tomorrow). Nothing is complete, but the starting blocks have
been laid for everything. We're going to need to modify our proposal, we
have to write the work schedule, we'll need to modify the current
progress file because right now it's just a copy of our project, and so
on. This change is just bringing in/creating the files we will need for
our submission, not writing anything into them or making them ready to
submit. The next couple of changes I will make are going to add actual
substance.

In addition to all the files I've already added, we need to determine
how we want to do the "signed commitment forms" required in the
instructions for the report. We can write, print and then sign them, or
we can just do them by hand, etc.. Unless you have a scanner though,
I'll need to take pictures of both forms and upload them as JPEGs. So
keep that in mind also.