Added Description
I've added the description of the code we have so far in a text cell.
There are probably a lot of spelling errors, but it gets the job done.

The instructions also suggest adding an explanation of what we've
learned about the subject matter for our project. I want to add
something to this effect, but I don't really know where to start.