It is Begun
So it turns out Dataset is a variable class in Mathematica, so I've
begun adding it into the project file and sectioning stuff off. We need
to start working on the functions for reading data into the interface.
Hopefully we can design both together, because if we wait to get this
database filled up  before we start to read it we'll be well into next
semester.