Data Consult Method
I've defined the syntax and parameters for what needs to be done with
the dataset access command ("dataSetConsult"). Basically it needs to be
given dataset parameters, determine whether the value exists in the
dataset, use that value if it exists and interpolate the value from
nearby ones if it doesn't. You can use the interpolate function you
wrote, but it may be easier to just incorporate that command into
dataSetConsult. I'm also going to search into the documentation center
to see if Mathematica has anything that can do this, and will help make
this function as it develops.