First Graphic Good
I've renamed the file, cause it needed to be done, and begun the whole
hooking things together process with what data I have.

Changes:
-name
-added real data to sphere graphic, using
"Quiet[Normal[Interpolation[All, column name]]]". This command setup
works and suppresses the error it gives when it has to extrapolate
something.

Non Permanent Changes:
-changed the range of the manipulate function from 0-1 to 0-0.5 because
that's as far as the database currently goes.

What I Need:
-I could use some help filling out the database to get complete star
data for the first graphic.
-The next step is to get the text readouts using real data, so you could
get started on that too.
-Also, I feel kinda dumb suggesting this when I jus spent an hour not
doing it, but we should start commenting this thing.