Adding Stuff
I changed the spreadsheet a little, added the text file from the
website, and started added some functionality to the text readouts and
sphere, changing the puny database I have to accommodate. I'm working
with the idea of large scale database functionality--making it as simple
as possible to switch datasets when we get the big one working. Thus,
I've put a lot of the unit conversions in the code, instead of
converting and then inputting the data into the dataset.

There's one new error: the ToString for core temp looks really weird,
probably because the number is so big. If you can solve it great,
otherwise we'll figure it out later.