I GOT IT DONE!
I wrote the code to convert the text file into a dataset. It's part of
the initialization process now. Also, this method does not convert
directly from text to dataset, it used table in the middle and lets us
change what data gets put into the dataset. So I'm converting data from
log forms, etc. in the transition, which takes some of the load off of
the main program. I am mega happy right now.