Death?
I took out the bit about not showing death sequences. that shouldn't be
too hard to do once we get the rest of the program done, though it will
be graphics intensive. The end result is directly dependent on the
staring mass, so it would basically be some graphics commands in a case
switch. We should be able to do it. I think. Probably. Maybe.