Changes
I made a little change that seems to make the interpolating functions
work out. I also added a little control in the interface to control the
starting age that will need to be tweaked. I really am sorry for not
working on this more. I really hope to be able to dedicate more time
while I'm in D.C.