Disks!
I just made the circles into disks. I think we might need to find a
better set of radii to work with for this. Also, I think that sliders to
bound the age slider should stay instead of integrating that into the
function call. Good job finding the option to make parameters optional!