HRD Plot Weirdness
We may want to meet with Claus about this. I've changed the parameters
to highlight the problem. The point on the HRD follows a different track
than the parametric plot, despite both using the exact same functions.
I've added an option into the plot to increase the quality, but the
point sampling is still weird. If there's no way to increase the quality
of the plot, then we may want to consider removing it.