Optimization and I Fixed It
So I wrote an okay solution to the plotting problem: by setting the
start and end values to the start and endpoints in the command, we can
get a more defined function with smaller timescales. it only kinda works
tho, and the downside is that with that smaller function range you only
see the plot of the iterating portion of the star's life.

Also, I added decimal points to the real numbers in our code, so that
they would be read as floating point, which is apparently faster (this
appears to be the case observationally). I'm gonna keep looking things
up to optimize our code, because the faster it runs, the more swag we
have.