YES!
I love the design, and have modified the text box a little to fit it, as
well as increase the length of the slider and window to match. However,
there are some things that need to be tweaked, such as text positioning
and how fast the larger slider updates. I'm going to try and meet with
Carini today about the core diagram also. If we cant get that working,
then we should probably stick to the old design because we'll have to
remove that. Also, the bar graph need tweaking if you wouldn't mind, the
key colors are all the same and I made a bad choice with the color
scaling that needs to be fixed.