Rid of a commented out piece of code
I just took out a piece of code that was commented out that we would not
be using. I am working on the interface. It (hopefully) should not take
that long. However, finding out how to put a background image on the
palettes may be difficult (I've already done a lot of searching).