Done!!!!!! (I think)
Everything seems to be operational. I added in demo spots that will be
referred to in the report (spots where the individual functions can
strut their stuff). All that you could do on this (if you so wish) is
find out how to stop the error messages between bounds and initialize.
Even then, I put a comment in there explaining why Quiet is used and why
that doesn't matter (it is commented by initialize). Thank you so much
for all the work that you have put into this project! I know much of the
work fell to you, and I hope that I am making a decent attempt at
compensating in this last stretch.