Alteration to time slider
I made the time slider longer because I think it looks better that way.
Make sure that if you run the code tomorrow (which you probably will)
that you delete all the output cells for any of the demo spots, so they
won't get a "Dynamic content" warning when they open the code.