Final Report Changes Part 1
I have put what I believe is a very nearly complete (if not complete)
part 1 to the report. Please look out for anything that you could make
better (as you do have a way with words). I also changed the "Output"
section in the code to be a "Demos" section where all the demo spots are
now located.